display .cody/ignore status in status bar (#2984)
Update status bar background color on ignored files (behind unstable
flag)

- Added functionality to the status bar in StatusBar.ts to display
whether the current file is ignored by Cody or not by switching the
background color of the status bar item.

This feature allows users and devs to easily identify if the current
file is being ignored by Cody directly from the editor.

This feature is behind the unstable feature flag. When unstable flag is
not enabled, the background will not be updated based on ignore file.

Note: I found it hard to tell if cody ignore is active or not during
development and testing. We can remove it when this is out of unstable
state.


## Test plan

<!-- Required. See
https://sourcegraph.com/docs/dev/background-information/testing_principles.
-->

1. Go to a Cody ignored file
2. Check if the status bar item update accordingly

### Updated

As discussed with Tim, we willl update the background color to indicates
if the current file is ignored by Cody or not instead of adding Ignore
next to Cody icon


![image](https://github.com/sourcegraph/cody/assets/68532117/2edb356f-201b-46b5-852a-1bf21a6cf2bd)

### Outdated


![image](https://github.com/sourcegraph/cody/assets/68532117/2294b1ba-3cc1-485c-b159-1bdf78916df7)