Agent: add support for `workspace/edit` (#2978)
Previously, the agent would crash when we hit on a codepath that used
the `vscode.WorkspaceEdit` class. This class is used to perform more
advanced edits on the workspace (multi-file, create files, rename files,
delete files, ..). This functionality is used by some inline-edit
commands like the `unit-tests` command that writes the generated test
case to the file instead of to the chat.

Now, the agent implements `vscode.WorkspaceEdit` so that it no longer
crashes. There are also several new JSON-RPC endpoints to support
applying workspace edits:

- `workspace/edit`: request sent from server to client to apply an edit
across
  multiple files.
- `workspace/didCreateFiles`: notification sent from client to server to
  notify files were created.
- `workspace/didDeleteFiles`: notification sent from client to server to
  notify files were deleted.
- `workspace/didRenameFiles`: notification sent from client to server to
  notify files were renamed.
- `textDocument/openUntitledDocument`: request sent from server to
client to open an untitled document.
- `textDocument/show`: request sent from server to client to prominently
show or make a document visible in the IDE.

The current implementation is not perfect. I didn't spend much time
perfecting all the details of how we apply edits. The goal of this PR is
to unblock further progress.

## Test plan

This PR adds basic tests for the new inline-edit test command that uses
`WorkspaceEdit`. I'm not super confident that the test client is
covering all cases. However, it's more important for IDE clients like
JetBrains to implement this endpoint 100% correctly.

<!-- Required. See
https://sourcegraph.com/docs/dev/background-information/testing_principles.
-->