Add config feature gating for chat and commands (#3039)
This commit adds a check for the chat and commands config features
before executing the chat command. If either feature is disabled by the
site admin, an error message is shown and the command is not executed.
This gates usage of the chat and commands behind the config features.

Also updated behavior where running a command in an ignored file will
returns an error.

## Test plan

<!-- Required. See
https://sourcegraph.com/docs/dev/background-information/testing_principles.
-->

connect to an instance that has chat or command disabled, and check if
the error message shows up correctly.