update: remove description requirement from custom command (#3025)
Remove the requirement to add a description for a custom command as the
prompt of the command is descriptive enough.

This eliminates one unnecessary step to simplify the custom command
creation process.

Based on screenshot shared by our users, it looks like users often use
the name of the command to fill in the description field, e.g.
https://sourcegraph.slack.com/archives/C04MSD3DP5L/p1706661347779839?thread_ts=1706656724.966049&cid=C04MSD3DP5L,
https://discord.com/channels/969688426372825169/1202562648340701257/1202562648340701257,
https://sourcegraph.slack.com/archives/C05MW2TMYAV/p1703011617179849

Feedback from a power user:

> One piece of feedback when defining your custom commands is to get rid
of the command description step.

## Test plan

<!-- Required. See
https://sourcegraph.com/docs/dev/background-information/testing_principles.
-->

We will still show the description in the command menu when provided,
but it is no longer a required field when creating a command from the
command palette. You can try to create a custom command with and without
the "description" field to confirm.

I've updated the current e2e tests and unit tests for custom commands to
reflect this change.