Make agent restart code much less error-prone (#566)
Fixes https://github.com/sourcegraph/jetbrains/issues/439

##
Changes
* Removed agent recovery loop, recovery is now on demand
* Added a bit of defensive code in case of unexpected exceptions
* Added few timeouts in places where we shouldn't wait forever
* Improved robustness of the shutdown
* Added additional safeguard for old agent binary file deletion

## Test plan
I added Thread that was destroying agent process in a fixed time after
startup.
I tested durations from 100ms to 10s.
Crashes after 200ms (which simulated complete agent failure) were not
spamming the logs with too much info, and were not leaving trash on the
disk.
Delay around 4-5s was enough for the chat to recover and give really
short answer. It was able to recover infinite amount of times, which
means we could survive even repeated crashes or hangs of the agent (if
that would ever be the case).