Autocomplete: Only sample requests that are also suggested (#3139)
Thinking about #3136 again, I don't think it's helpful to sample these
traces even. It's not bad per-se, but it's easier to aggregate if we
don't have to filter based on a root span attribute everywhere.

E.g. I want to see the "median" completion time so I get the median
timings of specific spans but if I have to exclude some values in the
root span, this is getting more complex then necessary. Let's keep it
simple and don't even sample these (which also saves some bandwidth)

## Test plan

- Added a debugger breakpoint to see if the sampled attribute is set