Command Hints: Update setting enablement (#3154)
## Description

1. Defaults the setting to on, as visible to the user. See comment below
as to why that is important
2. Adds default text to the setting description, so users can easily
search VS Code settings to disable this if they want to.

## Test plan

**When `cody.commandHints.enabled` setting is not set:**

1. Set `cody-command-hints` flag to 100%. Check command hints are
enabled. Check value is correctly shown in "Cody Settings" menu.
2. Set `cody-command-hints` flag to 0%. Reload and check command hints
are NOT enabled. Check value is correctly shown in "Cody Settings" menu.

**When `cody.commandHints.enabled` setting is already set to `true`:**

1. Set `cody-command-hints` flag to 0%. Reload and check command hints
are STILL enabled. Check value is correctly shown in "Cody Settings"
menu.

**When `cody.commandHints.enabled` setting is already set to `false`:**

1. Set `cody-command-hints` flag to 100%. Check command hints are NOT
enabled. Check value is correctly shown in "Cody Settings" menu.


<!-- Required. See
https://sourcegraph.com/docs/dev/background-information/testing_principles.
-->