Agent: make tests more stable
Previously, I was failing to run the agent tests locally in replay mode
from the latest `main` even if the tests were passing in CI. The root
cause was that we are not sorting context items that we feed to the
prompt. The CI tests seem to be flaky, because some open PRs are
reproducing the failure even if those PRs have unrelated changes.

This PR fixes the problem by adding an additional sorting step for
context items so that the tests pass reliably in replay mode.