Always uses latest stable VS Code version for integration tests (#3021)
This was previously hard-coded and needed updating (and confusingly in
two places, with one being bumped without the other in the past). This
just tracks latest stable.

If we want to also run tests on the oldest supported version (if this is
defined), we could split this out into an env VAR and have them both
(`['stable', 'x.y.z']`) in the GH Actions yaml file.

## Test plan

Non-functional... ensure bots are still running and the logs show that
they used the current latest version of VS Code (1.86).