Agent: correctly dispose code lens providers
Previously, we never respected the `dispose()` callback for code lens
providers due to a bug in the implementation. I haven't confirmed
100%, but this bug could have caused code lenses to accumulate for agent
clients. This PR fixes the bug where we handle the `dispose()` callback,
but I don't have time right now to add test cases or validate it fixes
the accumulation issue.