Remove excessive memory allocations during Chat History sorting (#784)
Related to #509

This simple fix introduces cache for ChatState.updatedAt date. The
getUpdatedTimeAt getter was causing memory allocations because each call
invoked the LocalDateTime.parse method. Value updatedAt is stored
internally as string. This date is used as a basis for sorting, so it's
intensively used during the Chat History tree update (in multiple
places).

## Test Plan

1. Start new conversation -> date in is properly initialized (today's
date) and saved to XML
2. Continue old conversation -> date is properly updated, moved to top
of the list, and saved to XML