rename type Context{File => Item}, remove existing redundant ContextItem type
These 2 types (ContextFile and ContextItem) were basically the same, except for a different field name (`content` vs. `text`) and a different default value (plus some other meaningless differences).

Also, ContextFile was a misnomer as this was being used for symbols as well (not just files).

There is no behavior change from this diff.