Update tokens incrementally in Chat window (#804)
Related to #509
Fixes #886

This commit aims to reduce CPU usage when using the Chat window when
Cody is responding.

Changes:

1. Bug #886 is fixed.
2. The chat is now updated only when the number of tokens grows. It
sometimes happens that Cody responds few times with the identical text
(in streaks, even 2-3 times) or responds in a different order. Such
responses are now ignored, so the panel is not unnecessarily refreshed.

## Test Plan

* Ask a question that will display a code block (e.g. "write hello world
in 10 different languages").
* The chat looks clean and does not have excessive paragraphs with "``"
symbols.