Update code formatting and sanitization in generic prompts and utils
This commit updates the code formatting and sanitization in the generic prompts and utils files.

In the generic prompts file, the system prompt for the 'edit' and 'add' intents has been updated to include a new rule for using correct indentations recommended by PEP 8 for Python code.

In the utils file, the contentSanitizer function has been updated to replace backslashes followed by an underscore with an underscore. This is to prevent any issues with markdown formatting.