preserve trailing non-alphanum chars in chat message
Given a chat message like `Explain @my/file.go :)`, all but the last non-alphanum char was dropped, so the resulting message as seen in the transcript would be `Explain @my/file.go)`. The handling of trailing non-alphanum was to not get tripped up by question marks, but it incorrectly assumed there would only be a single such character. This approach can be fixed more robustly by using a rich editor (https://github.com/sourcegraph/cody/pull/3287), but this is a quick fix for now.