disable model selector after 1st chat message is sent
In https://github.com/sourcegraph/cody/pull/3341 the behavior changed so that the welcome message goes away when you chat. (Which I love!) But there's a minor bug in that the model selector remains enabled after the 1st message is sent now (because the existing logic assumed that the welcome message was always the 1st message and so checked for `>1` messages to disable the model selector).