compute ChatMessage.displayText on-demand instead of persisting
Previously, `ChatMessage.displayText?: string` was set in multiple places, and the logic for getting the title of a chat was defined in 2 different places (and uses the `displayText`). Now, we never persist `displayText` and instead compute it at display time, which is possible because it uses only data that is available in the `ChatMessage` value.