prevent Cody from crashing on init with mis-serialized chat history
Similar to https://github.com/sourcegraph/cody/pull/3367, this prevents Cody from crashing on initialization if the chat history contains `vscode.Range` JSON-serialized values (which are JSON-serialized as `[start, end]`). This is an easy and subtle bug to introduce, and it's valuable to protect against complete crashing on initialization.

The stack trace of the crash that this prevents is:

```
TypeError: Cannot read properties of undefined (reading 'line')
	at createDisplayTextWithFileLinks (/home/sqs/.vscode/extensions/sourcegraph.cody-ai-1.8.2/dist/extension.node.js:124081:56)
	at getDisplayText (/home/sqs/.vscode/extensions/sourcegraph.cody-ai-1.8.2/dist/extension.node.js:124292:12)
	at prepareChatMessage (/home/sqs/.vscode/extensions/sourcegraph.cody-ai-1.8.2/dist/extension.node.js:124279:18)
	at groupCodyChats (/home/sqs/.vscode/extensions/sourcegraph.cody-ai-1.8.2/dist/extension.node.js:124336:26)
...
```