two bug fixes for inline edits
Fixed a serious bug with document contents disappearing in the Agent.  The problem was that textDocument/didFocus would simply move the document from the wire into the workspace's document map for that URI. JetBrains was not sending the document content along with every didFocus, because it clutters the logs significantly, and the contents have not changed.

I fixed this on the Agent side by copying in the existing document's content if none was passed.

Also added 'editTask/getFoldingRanges' request to expose the indentation-based folding ranges API to clients (such as JetBrains) which cannot reliably compute folding ranges in a language-neutral way. This fixed a bug with not knowing the initial position for the spinner while we wait for the FixupController to reply to our initial commands/document request.