support @-mentioning URLs to use web page contents as context
This is an experimental (off-by-default) feature that lets you, in chat, @-mention web pages by URL to include their text contents as context for the LLM. For example, you can type `what are the books listed at @https://slack.org/reading` into chat, select the context item in the menu, and then press Enter, and Cody will read the contents of that web page.

Requires the experimental and undocumented `"cody.experimental.urlContext": true` VS Code setting.

This is just something we can experiment with internally for now.