only allow 1 retry for flaky e2e tests
This will mean more annoying e2e failures due to flakiness in our PRs, which is bad. But it, along with https://github.com/sourcegraph/cody/pull/3446, will mean that we (if we are good people :) will notice the flakiness better and will fix it, thereby making it so we don't need to wait for multiple retries of some flaky tests on each build, thereby saving us time overall.