remove `typing @-mention text does not automatically accept it` e2e test
This e2e test no longer makes sense now that we changed the behavior so that pressing <kbd>Space</kbd> actually does accept the first suggestion. We could add a test for pressing something else like <kbd>,</kbd>, but that seems of marginal importance and we might even want <kbd>,</kbd> and other punctuation to select the suggestion in the future. The reason this test was passing is due to flakiness, a race condition where zero context items were shown because the response had not loaded yet, not because the behavior was as expected.