use vscode-jsonrpc library for JSON-RPC implementation
The [vscode-jsonrpc](https://www.npmjs.com/package/vscode-jsonrpc) library is from the VS Code team and is used by many VS Code extensions and language servers to implement the JSON-RPC 2.0 protocol. It is well tested, heavily used, actively maintained, and supports both Node and Web platforms.

Using this library means we can remove our own implementation of the protocol, which is (a) less code to maintain, (b) does not support the browser, and (c) is not spec-compliant (for example, it supports string values as request parameters, which are disallowed by the [JSON-RPC 2.0 spec](https://www.jsonrpc.org/specification#request_object)).