Fix bug where up/down arrow keys always navigate chat history (#1323)
Fixes #1134

Previously, pressing up/down arrow keys in the chat input always
replaced the entire contents of the chat input with the previous/next
chat message from history. This was a regression that was introduced in
the commit
https://sourcegraph.com/github.com/sourcegraph/jetbrains/-/commit/88a21694097716fbe53b84e76b3124cb3454af76#diff-91c99de25d8f01dab68bee8fa28fb6f1L80
where the variable `isInHistoryMode` got removed. This PR adds back the
original logic for `isInHistoryMode` so the chat input works like
before.

## Test plan

Manually tested. See demo
https://www.loom.com/share/4806fcaeaa9e44519d12e8b6d16a29a0
<!-- All pull requests REQUIRE a test plan:
https://sourcegraph.com/docs/dev/background-information/testing_principles

Why does it matter?

These test plans are there to demonstrate that are following industry
standards which are important or critical for our customers.
They might be read by customers or an auditor. There are meant be simple
and easy to read. Simply explain what you did to ensure
your changes are correct!

Here are a non exhaustive list of test plan examples to help you:

- Making changes on a given feature or component:
- "Covered by existing tests" or "CI" for the shortest possible plan if
there is zero ambiguity
  - "Added new tests"
- "Manually tested" (if non trivial, share some output, logs, or
screenshot)
- Updating docs:
  - "previewed locally"
  - share a screenshot if you want to be thorough
- Updating deps, that would typically fail immediately in CI if
incorrect
  - "CI"
  - "locally tested"
-->