Graph Context: add `tsc` strategy.
This PR adds a new autocomplete graph context strategy `tsc-mixed` that
uses the TypeScript compiler (aka. `tsc`) directly to figure out what
symbols and types to add to the Cody context. In comparison to LSP, we
get access to richer APIs with tsc directly. For example, with tsc,
we can implement custom formatting for complicated types so that the LLM
is more likely to generate valid code.