improve contrast of context links and at-mention colors in various VS Code themes
Some VS Code color themes made the context links and at-mentions in chat unreadable. This was because the styling assumed that the text link foreground had contrast with the list active selection background, which is an incorrect assumption in general. Now the styles are formally shared between these 2 components and use CSS theme variables that can be assumed to have appropriate contrast.

Fix https://github.com/sourcegraph/cody/issues/3816