QA: add manual testing step for up/down arrow keys (#1336)
Previously, we didn't document the manual steps to test the behavior of
up/down arrow keys in the chat input. We experienced a regression in
this functionality around February. This PR adds the manual testing
steps to prevent a similar regression from happening again.

## Test plan

Included in testing.md
<!-- All pull requests REQUIRE a test plan:
https://sourcegraph.com/docs/dev/background-information/testing_principles

Why does it matter?

These test plans are there to demonstrate that are following industry
standards which are important or critical for our customers.
They might be read by customers or an auditor. There are meant be simple
and easy to read. Simply explain what you did to ensure
your changes are correct!

Here are a non exhaustive list of test plan examples to help you:

- Making changes on a given feature or component:
- "Covered by existing tests" or "CI" for the shortest possible plan if
there is zero ambiguity
  - "Added new tests"
- "Manually tested" (if non trivial, share some output, logs, or
screenshot)
- Updating docs:
  - "previewed locally"
  - share a screenshot if you want to be thorough
- Updating deps, that would typically fail immediately in CI if
incorrect
  - "CI"
  - "locally tested"
-->

---------

Co-authored-by: Daniel Marques <daniel.marques@sourcegraph.com>