Agent: add `server` subcommand to expose agent protocol via websockets
Previously, it was only possible to do RPC with the Cody Agent via
stdin/stdout, which required the client to run on the same computer as
the Cody Agent. This PR adds a new `server` subcommand that starts a
server that allows clients to connect with the Cody Agent via web
sockets from another computer.

One example use-case for this feature is to implement a browser client
to interact with Cody Chat. By using websockets, the browser client can
interact with Cody without installing any custom dependencies. To
improve the ergonomics of interacting with Cody, we can extend our
current Kotlin bindings generation tool to generate bindings for other
languages like TypeScript and Go.