Display onboarding tooltip on first autocomplete suggestion (#1339)
Fixes #1319.

Previously, users didn't get any onboarding instructions when using Cody
for the first time in JetBrains. For example, many users didn't know
they should press TAB to accept an autocomplete suggestion, or
option-bracket to cycle through options. This PR fixes the problem by
using the built-in `GotItTooltip` component, which is designed for this
use-case.

## Test plan

Manually tested. 
![CleanShot 2024-04-18 at 16 01
48@2x](https://github.com/sourcegraph/jetbrains/assets/1408093/5d01db65-72e2-44a1-8288-8dbac5ea6ab8)


<!-- All pull requests REQUIRE a test plan:
https://sourcegraph.com/docs/dev/background-information/testing_principles

Why does it matter?

These test plans are there to demonstrate that are following industry
standards which are important or critical for our customers.
They might be read by customers or an auditor. There are meant be simple
and easy to read. Simply explain what you did to ensure
your changes are correct!

Here are a non exhaustive list of test plan examples to help you:

- Making changes on a given feature or component:
- "Covered by existing tests" or "CI" for the shortest possible plan if
there is zero ambiguity
  - "Added new tests"
- "Manually tested" (if non trivial, share some output, logs, or
screenshot)
- Updating docs:
  - "previewed locally"
  - share a screenshot if you want to be thorough
- Updating deps, that would typically fail immediately in CI if
incorrect
  - "CI"
  - "locally tested"
-->