models: show default LLM first in dropdown (#3859)
The order we list our models in dotcom.ts is the order it is shown in. I imagine in future we likely want to order these such that LLMs we want to guide users to use are higher up. For now I just moved the default LLM to the top.

Test Plan: CI