remove misleading `| undefined` return for editor file content
`Editor.getTextEditorContentForFile` now returns `Promise<string>` not `Promise<string | undefined>`. The only case where the impl returned an `undefined` content value was if the URI argument was undefined, which was impossible per the type signature anyway and does not make sense.

Now, reading a file either returns a string or throws an exception, which makes sense. It does not make sense for it to ever return `undefined`.