do not show the context file line count in chat context
Previously the chat context cell would display `Context: 93 lines from 7 files`. The number of lines is not necessary, and now it just shows the file count, as in `Context: 7 files`.

The total line count is an unnecessary detail even for today's context sources. It is not clear what users should do with that information. As context sources start to get less about "chunks of files", it becomes even less relevant.