feat(ignore-helper): skip ignoring http URIs
The IgnoreHelper class now checks if the URI scheme is 'http' and returns false (i.e., does not ignore the URI) in that case. This change allows the code to process HTTP URIs, which were previously being ignored along with other non-file URIs.

The updated logic in the shouldIgnore method is:

- If the URI scheme is 'http', return false (do not ignore)

This change ensures that HTTP URIs are handled correctly by the application, potentially enabling new functionality or fixing an issue related to HTTP resource handling.