test and support node 18 (no longer node 16)
My recommendation is that we should stop testing and supporting Node 16, and instead bump this to Node 18.

We had been supporting Node 16 because the July 2023 VS Code release shipped with a version of Electron that used Node 16. That is now well in the past, and:

- VS Code's Aug 2023 release shipped with Node 18 (up from Node 16): https://code.visualstudio.com/updates/v1_82#_engineering
- Node 16 was end-of-life'd on 2023-09-11: https://nodejs.org/en/blog/announcements/nodejs16-eol
- For JetBrains, we use the agent binary that bundles Node (20), so we don't depend on Node 16 even if that's what is installed on the system.

An alternative is to start testing Node 16, 18, AND 20 (ie add tests for Node 18 since it's the version in use by VS Code). This would be safer in case any customers are still using an older version of VS Code that requires Node 16. I'll confirm this before merging.